deployables/ai_agent_env/agents/file_processor.py
import functions_framework
from google.cloud import storage
from google.cloud import aiplatform
import pandas as pd
import io
import json

# Initialize Vertex AI
aiplatform.init(project='your-project-id', location='us-central1')

@functions_framework.cloud_event
def validate_csv_ai(cloud_event):
    data = cloud_event.data
    bucket_name = data["bucket"]
    file_name = data["name"]

    # Initialize the Cloud Storage client
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(file_name)

    # Download the file content
    content = blob.download_as_text()

    # Convert CSV to DataFrame
    df = pd.read_csv(io.StringIO(content))

    # Perform AI-driven validation
    validation_result = ai_validate_data(df)

    # Store the validation result
    result_blob = bucket.blob(f"{file_name}_ai_validation_result.json")
    result_blob.upload_from_string(json.dumps(validation_result))

    print(f"AI Validation complete for {file_name}. Result stored in {file_name}_ai_validation_result.json")

def ai_validate_data(df):
    anomalies = []
    
    # Use Vertex AI for anomaly detection
    endpoint = aiplatform.Endpoint('your-vertex-ai-endpoint')
    
    # Prepare data for the model
    instances = df.to_dict(orient='records')
    
    # Get predictions from the model
    predictions = endpoint.predict(instances=instances)
    
    # Process predictions to identify anomalies
    for i, prediction in enumerate(predictions.predictions):
        if prediction['anomaly_score'] > 0.8:  # Threshold for anomaly
            anomalies.append({
                'row': i,
                'score': prediction['anomaly_score'],
                'features': prediction['feature_importance']
            })

    return {
        "file_valid": len(anomalies) == 0,
        "row_count": len(df),
        "anomaly_count": len(anomalies),
        "anomalies": anomalies
    }